Refactored send_state to prepare for updateWorkerPilotStatus
